TST: Test import cleanups
